+++ samplesheets/src/components/Overview.vue
@@ -23,6 +23,10 @@
           <dd class="col-md-9">{{ app.sodarContext['configuration']}}</dd>
         </dl>
         <dl class="row pb-0">
+          <dt class="col-md-3">Parser Version</dt>
+          <dd class="col-md-9">{{ app.sodarContext['parser_version']}}</dd>
+        </dl>
+        <dl class="row pb-0">
           <dt class="col-md-3">iRODS Repository</dt>
           <dd class="col-md-9">
             <span v-if="app.sodarContext['irods_status']"